Title: 최솟값 3개, Time: 229ms, Memory: 1MB, Status: Passed - CodeTree

+++ 231227/최솟값 3개/top-3-smallest-number.cpp
@@ -21,7 +21,7 @@ int main() {
             pq.pop();
             int c=-pq.top();
             pq.pop();
-            cout<<a*b*c<<'\n';
+            cout<<(long long)a*b*c<<'\n';
 
             pq.push(-a);
             pq.push(-b);
